requirements.txt
bleak>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
sendgrid>=6.9.0 